+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ------------------
 
   * added `bovigo\callmap\verify()` as possibility to verify an expected call amount
+  * retrieving call for a method which doesn't exist or is not applicable for mapping now throws an `\InvalidArgumentException`
+  * retrieving received arguments for a method which doesn't exist or is not applicable for mapping now throws an `\InvalidArgumentException`
 
 
 0.4.0 (2015-04-12)